requirements.txt.txt
Flask==3.0.0
Pillow==10.2.0
numpy==1.26.0
qrcode==7.4.2
gunicorn==21.2.0
werkzeug==3.0.1